// src/lib/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  public: {
    Tables: {
// ...
      sections: {
        Row: {
          chapter_id: string
          content_html: string | null
          content_json: Json | null
          created_at: string
          id: string
          serial_number: number
          sub_title: string | null
          title: string
          updated_at: string
        }
        Insert: {
          chapter_id: string
          content_html?: string | null
          content_json?: Json | null
          created_at: string
          id?: string
          serial_number: number
          sub_title?: string | null
          title: string
          updated_at: string
        }
        Update: {
          chapter_id?: string
          content_html?: string | null
          content_json?: Json | null
          created_at?: string
          id?: string
          serial_number?: number
          sub_title?: string | null
          title?: string
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "sections_chapter_id_fkey"
            columns: ["chapter_id"]
            isOneToOne: false
            referencedRelation: "chapters"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}